Move BaseRenderLayer into browser
